style the form
