+++ src/App.js
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
-import Books from './components/Book';
+import Books from './components/Books';
 import Categories from './components/Categories';
 import Navbar from './components/Navbar';
 import NotFound from './components/NotFound';
